# Make map 3.5-4.8 times faster