## Julia 1.5 ~4.8x
### Without this PR
```julia
julia> @btime (x->last(map(sin, fill(x,100))))'(0.1)
  15.635 μs (769 allocations: 23.53 KiB)
```

### With this PR
```julia
julia> @btime (x->last(map(sin, fill(x,100))))'(0.1)
  3.182 μs (35 allocations: 8.23 KiB)
```

## Julia 1.4 ~3.5x faster
### Without this PR
```julia
julia> @btime (x->last(map(sin, fill(x,100))))'(0.1)
  16.815 μs (769 allocations: 22.88 KiB)
0.9950041652780258
```

### With this PR
```julia
julia> @btime (x->last(map(sin, fill(x,100))))'(0.1)
  4.751 μs (57 allocations: 8.14 KiB)
````