@dhairyagandhi96 2020-05-21
Failure to build docs because of probably wants to be run on julia 1.3+