@cossio 2020-05-21
Would it make sense to generalize this to:

```julia
@generated function unzip(tuples::AbstractVector{<:NTuple{N}}) where {N}
  Expr(:tuple, (:(getindex.(tuples, $i)) for i ∈ 1:N)...)
end
``